split rate limit test to a new `gateway.rs` file`
